feat(engine): extract headers into user task custom headers
Process task headers provided for a user task element into "customHeaders" of user task records.
The HeaderEncoder is extracted from the BpmnJobBehavior into a dedicated public class for reuse.
For Zeebe user tasks, the task headers are encoded like job headers are. No additional headers
are added by the engine.
